requirements.txt
Package                       Version
----------------------------- ---------
asgiref                       3.5.2
asyncio                       3.4.3
beautifulsoup4                4.12.2
bs4                           0.0.1
certifi                       2021.10.8
charset-normalizer            2.0.9
cloudinary                    1.30.0
coreapi                       2.3.3
coreschema                    0.0.4
Django                        3.2.16
django-cors-headers           3.13.0
django-decouple               2.1
django-phonenumber-field      7.0.2
django-phonenumbers           1.0.1
djangorestframework           3.13.1
djangorestframework-simplejwt 5.0.0
drf-yasg                      1.20.0
greenlet                      2.0.1
gunicorn                      20.1.0
idna                          3.3
importlib-metadata            6.0.0
inflection                    0.5.1
itypes                        1.2.0
Jinja2                        3.0.3
MarkupSafe                    2.0.1
packaging                     21.3
phonenumbers                  8.13.6
pip                           22.3.1
PyJWT                         2.6.0
pyotp                         2.8.0
pyparsing                     3.0.6
pytz                          2021.3
requests                      2.26.0
ruamel.yaml                   0.17.19
ruamel.yaml.clib              0.2.6
setuptools                    65.5.0
six                           1.16.0
soupsieve                     2.5
sqlparse                      0.4.2
typing_extensions             4.4.0
uritemplate                   4.1.1
urllib3                       1.26.7
zipp                          3.12.0
